requirements.txt
# Autopilot Pro - Python Dependencies
# =====================================
# Install all dependencies with: pip install -r requirements.txt

# Core ML and Computer Vision
ultralytics>=8.0.0          # YOLO models
opencv-python>=4.8.0        # Image processing
numpy>=1.24.0               # Numerical operations
Pillow>=10.0.0              # Image handling

# Web Interface
gradio>=4.0.0               # Gradio web interface

# HTTP Requests
requests>=2.31.0            # For server health checks

# Optional but recommended
torch>=2.0.0                # PyTorch (required for YOLO)
torchvision>=0.15.0         # Vision utilities

